pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "docx>=0.2.4",
    "openai>=1.88.0",
    "pdfplumber>=0.11.7",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.46.0",
]
